+++ libs/core/src/lib/multi-input/multi-input.component.ts
@@ -38,14 +38,15 @@ import {
     FocusEscapeDirection,
     KeyUtil,
     FocusTrapService,
-    uuidv4
+    uuidv4,
+    RangeSelector
 } from '@fundamental-ngx/core/utils';
 
 import { MultiInputMobileComponent } from './multi-input-mobile/multi-input-mobile.component';
 import { MultiInputMobileModule } from './multi-input-mobile/multi-input-mobile.module';
 import { MULTI_INPUT_COMPONENT, MultiInputInterface } from './multi-input.interface';
 import { SelectionModel } from '@angular/cdk/collections';
-import { map, startWith } from 'rxjs/operators';
+import { distinctUntilChanged, first, map, shareReplay, startWith } from 'rxjs/operators';
 import { uniqBy } from 'lodash-es';
 
 /**
@@ -313,6 +314,9 @@ export class MultiInputComponent
     private _subscriptions = new Subscription();
 
     /** @hidden */
+    private readonly _rangeSelector = new RangeSelector();
+
+    /** @hidden */
     onChange: (value: any) => void = () => {};
 
     /** @hidden */
@@ -356,8 +360,10 @@ export class MultiInputComponent
         }
 
         this._subscriptions.add(
-            this._searchTermCtrl.valueChanges.subscribe((searchTerm) => {
+            this._searchTermCtrl.valueChanges.pipe(distinctUntilChanged()).subscribe((searchTerm) => {
                 this.searchTermChange.emit(searchTerm);
+                // resetting existing selection state, if any
+                this._rangeSelector.reset();
             })
         );
     }
@@ -488,11 +494,43 @@ export class MultiInputComponent
     }
 
     /** @hidden */
-    _handleSelect(checked: any, value: any, resetSearch = true, event?: MouseEvent): void {
-        if (event) {
-            event.preventDefault(); // prevent this function from being called twice when checkbox updates
+    _onCheckboxKeyup(value: any, event: KeyboardEvent, index: number): void {
+        if (KeyUtil.isKeyCode(event, [SPACE, ENTER])) {
+            this._onCheckboxClick(value, event, index);
         }
+    }
 
+    /** @hidden */
+    async _onCheckboxClick(
+        value: any,
+        event: PointerEvent | KeyboardEvent,
+        index: number,
+        isListItem = false
+    ): Promise<void> {
+        const toggledSelection = !this._selectionModel.isSelected(value);
+        this._rangeSelector.onRangeElementToggled(index, event);
+        const sub = this.viewModel$.pipe(first()).subscribe((vm) => {
+            this._rangeSelector.applyValueToEachInRange((idx) =>
+                this._handleSelect(toggledSelection, vm.displayedOptions[idx].value, false)
+            );
+        });
+        this._subscriptions.add(sub);
+        if (isListItem) {
+            this.openChangeHandle(false);
+        } else {
+            // stop propagation on the checkbox so event doesn't reach the list item
+            event.stopPropagation();
+        }
+    }
+
+    /** @hidden */
+    _onTokenClick(value: any, resetSearch: boolean, event?: PointerEvent): void {
+        event?.preventDefault(); // prevent this function from being called twice when checkbox updates
+        this._handleSelect(false, value, resetSearch);
+    }
+
+    /** @hidden */
+    _handleSelect(checked: any, value: any, resetSearch = true): void {
         const previousLength = this._selectionModel.selected.length;
         if (checked) {
             this._selectionModel.select(value);
@@ -515,7 +553,6 @@ export class MultiInputComponent
         this._propagateChange();
     }
 
-    /** @hidden */
     _handleInputKeydown(event: KeyboardEvent): void {
         if (KeyUtil.isKeyCode(event, DOWN_ARROW) && !this.mobile) {
             if (event.altKey) {
@@ -559,13 +596,23 @@ export class MultiInputComponent
 
     /** @hidden */
     _onSubmit(): void {
-        if (this.allowNewTokens && this.newTokenValidateFn(this._searchTermCtrl.value)) {
+        const searchTerm = this.searchTerm;
+        if (searchTerm === '') {
+            return;
+        }
+        const isExist = this._selectFirstFiltered(searchTerm);
+        if (!isExist && this.allowNewTokens && this.newTokenValidateFn(this._searchTermCtrl.value)) {
             const newToken = this.newTokenParseFn(this._searchTermCtrl.value);
-            this.dropdownValues.push(newToken);
+            this._addNewTokenToDropDownValues(newToken);
             this._handleSelect(true, newToken);
             this._searchTermCtrl.setValue('');
             this.open = false;
         }
+    }
+
+    /** @hidden */
+    _handleComplete({ term }): void {
+        this.searchTerm = term;
     }
 
     /**
@@ -604,6 +651,25 @@ export class MultiInputComponent
     /** @hidden */
     enableParentFocusTrap(): void {
         this._focusTrapService?.unpauseCurrentFocusTrap();
+    }
+
+    /** @hidden */
+    private _addNewTokenToDropDownValues(newToken): void {
+        this.dropdownValues.push(newToken);
+        const newOption = this._getOptionItem(newToken);
+        this.optionItems$.next([...this.optionItems$.value, newOption]);
+    }
+
+    /** @hidden */
+    private _selectFirstFiltered(searchTerm: string): boolean {
+        const filtered = this.filterFn(this.dropdownValues, searchTerm);
+        if (Array.isArray(filtered) && filtered.length > 0 && !this.includes && this.autoComplete) {
+            this._handleSelect(true, filtered[0]);
+            this._searchTermCtrl.setValue('');
+            this.open = false;
+            return true;
+        }
+        return false;
     }
 
     /** @hidden */
@@ -728,7 +794,8 @@ export class MultiInputComponent
                     selectedOptions,
                     displayedOptions
                 };
-            })
+            }),
+            shareReplay({ refCount: true, bufferSize: 1 })
         );
     }
 }