+++ libs/platform/src/lib/form/public_api.ts
@@ -53,6 +53,7 @@ export * from './form-group/fdp-form.module';
 export * from './form-group/form-group.component';
 export * from './form-group/form-field/form-field.component';
 export * from './form-group/form-field-group/form-field-group.component';
+export * from './form-group/form-field-extras/form-field-extras.component';
 
 export * from './input/fdp-input.module';
 export * from './input/input.component';